requirements.txt
discord
openai
tiktoken
loguru
python-dotenv
revChatGPT